+++ src/pages/calendar/components/EventSearch/features/index.ts
@@ -1,0 +1,1 @@
+export * from './EventList';
