+++ src/pages/calendar/components/EventView/features/ViewNavigation.tsx
@@ -1,0 +1,34 @@
+import { ChevronLeftIcon, ChevronRightIcon } from '@chakra-ui/icons';
+import { HStack, IconButton } from '@chakra-ui/react';
+
+import { SelectField } from '@/shared/ui';
+
+interface ViewNavigationProps {
+  view: 'week' | 'month';
+  setView: (view: 'week' | 'month') => void;
+  navigate: (direction: 'prev' | 'next') => void;
+}
+
+export const ViewNavigation = ({ view, setView, navigate }: ViewNavigationProps) => {
+  return (
+    <HStack mx="auto" justifyContent="space-between">
+      <IconButton
+        aria-label="Previous"
+        icon={<ChevronLeftIcon />}
+        onClick={() => navigate('prev')}
+      />
+
+      <SelectField
+        label="View"
+        value={view}
+        onChange={(e) => setView(e.target.value as 'week' | 'month')}
+        options={[
+          { value: 'week', label: 'Week' },
+          { value: 'month', label: 'Month' },
+        ]}
+      />
+
+      <IconButton aria-label="Next" icon={<ChevronRightIcon />} onClick={() => navigate('next')} />
+    </HStack>
+  );
+};
