+++ src/pages/calendar/components/EventView/features/index.ts
@@ -1,0 +1,3 @@
+export * from './MonthView';
+export * from './WeekView';
+export * from './ViewNavigation';
